rafactored

+++ source/services/user.services.ts
@@ -9,7 +9,7 @@ export const fetchUsers = async (): Promise<User[] | any> => {
 }
 
 export const findUserById = async (userId: string): Promise<User | any> => {
-    const user: UserEntity = await userRepository.fetchUserById(userId)
+    const user = await userRepository.fetchUserById(userId)
     if (!user) {
         throw new Error(`User with ${userId} does not exist`)
     }
@@ -20,7 +20,7 @@ export const findUserById = async (userId: string): Promise<User | any> => {
 export const findUserByEmail = async (email: string): Promise<User | null> => {
     const user = await userRepository.fetchUserByEmail(email)
     if (!user) {
-        return null
+        throw new Error(`User ${email} does not exist`)
     }
     const response = mapUserFromUserEntity(user)
     return response
@@ -39,12 +39,19 @@ export const createUser = async (user: User): Promise<User> => {
 
 export const updateUser = async (userId: string, user: User): Promise<User> => {
     const userEntity = mapUserEntityFromUser(user)
+    if (!userEntity.username) {
+        throw new Error(`User with ${userId} does not exist`)
+    }
     const [db_response] = await userRepository.updateUser(userId, userEntity)
     const response = mapUserFromUserEntity(db_response)
     return response
 }
 
 export const deleteUser = async (userId: string): Promise<User> => {
+    const user = await userRepository.fetchUserById(userId)
+    if (!user) {
+        throw new Error(`User with ${userId} does not exist`)
+    }
     const [db_response] = await userRepository.deleteUser(userId)
     const response = mapUserFromUserEntity(db_response)
     return response
